# Don't use padding in base32 helpers