@tmthrgd 2018-05-21
I threw together a simple benchmark and the call to `WithPadding` has overhead:

```
name        old time/op    new time/op    delta
ToBase32-8     242ns ± 6%     501ns ± 7%  +106.80%  (p=0.000 n=10+8)

name        old alloc/op   new alloc/op   delta
ToBase32-8      128B ± 0%      416B ± 0%  +225.00%  (p=0.000 n=10+10)

name        old allocs/op  new allocs/op  delta
ToBase32-8      2.00 ± 0%      3.00 ± 0%   +50.00%  (p=0.000 n=10+10)
```

The performance overhead goes away entirely if you extract `base32.HexEncoding.WithPadding(base32.NoPadding)` into a global variable.